SFB-65: move items form /sections to /section
