+++ app/utils/code-editor/completion/section/display-types.js
