+++ app/utils/code-editor/completion/section/sh-predicates.js
